+++ examples/components/examples/index.js
@@ -1,0 +1,7 @@
+export Simple from './Simple';
+export Basic from './Basic';
+export Static from './Static';
+export Confirm from './Confirm';
+export BasicConfirm from './BasicConfirm';
+export Long from './Long';
+export Markdown from './Markdown';