+++ examples/main.js
@@ -1,6 +1,8 @@
 import React from 'react';
 
-import App from './app';
+import App from './components/App';
+
+import packageCfg from '../package.json';
 
 
-React.render(<App />, document.getElementById('content'))
+React.render(<App packageCfg={packageCfg} />, document.getElementById('content'))