+++ webpack.config.js
@@ -1,5 +1,21 @@
 var path = require('path');
 var webpack = require('webpack');
+var fs = require('fs');
+
+
+function collectExampleSource() {
+    var baseDir = path.join(__dirname, 'examples', 'components', 'examples');
+    var files = fs.readdirSync(baseDir);
+    var result = {};
+
+    files.forEach(function (fileName) {
+        if (/\.jsx?$/.test(fileName)) {
+            result[fileName.replace(/\.jsx?$/, '')] = fs.readFileSync(path.join(baseDir, fileName), {encoding: 'utf-8'});
+        }
+    });
+
+    return result;
+}
 
 
 module.exports = {
@@ -23,10 +39,16 @@ module.exports = {
     },
     plugins: [
         new webpack.HotModuleReplacementPlugin(),
-        new webpack.IgnorePlugin(/un~$/)
+        new webpack.IgnorePlugin(/un~$/),
+        new webpack.DefinePlugin({
+            EXAMPLE_SRC: JSON.stringify(collectExampleSource())
+        })
     ],
     resolve: {
-        extensions: ['', '.js']
+        extensions: ['', '.js'],
+        alias: {
+            bootstrap: path.join(__dirname, 'node_modules', 'bootstrap-sass', 'assets', 'stylesheets', 'bootstrap')
+        }
     },
     module: {
         loaders: [
@@ -36,8 +58,12 @@ module.exports = {
                 loader: 'babel'
             },
             {
-                test: /\.md?$/,
+                test: /\.md/,
                 loader: 'raw'
+            },
+            {
+                test: /\.json/,
+                loader: 'json'
             },
             {
                 test: /\.css/,
@@ -45,7 +71,8 @@ module.exports = {
             },
             {
                 test: /\.scss/,
-                loader: 'style!css!autoprefixer!compass'
+                loader: 'style!css!autoprefixer!sass?'+
+                        'includePaths[]=' + encodeURIComponent(path.resolve(__dirname, 'node_modules', 'bootstrap-sass', 'assets', 'stylesheets'))
             }
         ]
     }