+++ CHANGELOG.md
@@ -2,10 +2,19 @@
 <!-- DON'T EDIT THIS SECTION, INSTEAD RE-RUN doctoc TO UPDATE -->
 # Changelog
 
+- [v0.1.0](#v010)
 - [v0.0.1](#v001)
 
 <!-- END doctoc generated TOC please keep comment here to allow auto update -->
 
 ### v0.1.0
+
+  * Core logic
+  * Added examples
+  * Basic modal variant
+  * Static modal
+  * Handle keyboard events
+
+### v0.0.1
 
   * tg-modal initial commit.